Fix corner cases in Bash runfiles library
Due to tests not actually verifying the exit code of rlocation calls,
a few special cases were not handled correctly. In particular, absolute
paths still went through regular rlocation lookup and could fail there
even after the absolute path had been printed to stdout.

Also adds a missing newline in the (very rare) case that a manifest
entry is found, but doesn't point to an existing file (e.g. if it is an
unresolved symlink), and `rlocation` is not used with command
substitution, but some other mechanism that doesn't strip trailing
newlines.

The tests now assert the expected exit code (== 0 or != 0).

Fixes #16933